moving and renaming storage unit container and components
