+++ frontend/src/containers/storageunit/index.js
@@ -1,0 +1,7 @@
+import StorageUnitList from './grid'
+import StorageUnitContainer from './panel'
+
+export {
+  StorageUnitList,
+  StorageUnitContainer
+}
