+++ frontend/src/routes.js
@@ -21,10 +21,10 @@ import React from 'react'
 import { IndexRedirect, Route } from 'react-router'
 import NotFound from './components/NotFound'
 import WelcomeView from './containers/welcome-view'
-import StorageUnitsView from './containers/storageunits-view'
+import StorageUnitsTable from './containers/storageunit/grid'
 import PickListView from './containers/picklist'
 import ExampleView from './containers/example-view'
-import StorageUnitContainer from './containers/storageunit-container'
+import StorageUnitPanel from './containers/storageunit/panel'
 import WelcomeUserView from './containers/welcome-user'
 import ObservationView from './containers/observation'
 import App from './containers/app'
@@ -56,11 +56,11 @@ export default (store) => {
       <Route path="/" component={WelcomeView} onEnter={redirectIfLoggedIn} />
 
       <Route path="/picklist" component={PickListView} />
-      <Route path="/storageunits" component={StorageUnitsView} />
+      <Route path="/storageunits" component={StorageUnitsTable} />
       <Route path="/example" component={ExampleView} />
       <Route path="/observation" component={ObservationView} />
-      <Route path="/storageunit/:id" component={StorageUnitContainer} />
-      <Route path="/storageunit/add" component={StorageUnitContainer} />
+      <Route path="/storageunit/:id" component={StorageUnitPanel} />
+      <Route path="/storageunit/add" component={StorageUnitPanel} />
 
       -- Authentication routes
       <Route path="/musit" component={WelcomeUserView} onEnter={requireLogin} />
